+++ app/wireshell.php
@@ -49,7 +49,7 @@ if (file_exists(__DIR__.'/../../../autoload.php')) {
     require __DIR__.'/../vendor/autoload.php';
 }
 
-$app = new Application('wireshell - An extendable ProcessWire CLI', '1.0.3');
+$app = new Application('wireshell - An extendable ProcessWire CLI', '1.0.6');
 
 $app->add(new UserCreateCommand());
 $app->add(new UserUpdateCommand());